+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - mtr plugin py3 compatibility issues
 - fileprobe lines that can't be processed are now skipped
 - fix #62 timeseriesdb plugins crash when `update` is given `None` as a value
+- fix #54 pytest > 3 for testing
 ### Changed
 - move from `facsimile` to `ctl` for package / release management
 - parse_interval now can handle multiunit strings such as 1m30s